requirements.txt
# HERMES OS — Dependencies
# ─── Core Framework ──────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ─── Telegram Bot ────────────────────
python-telegram-bot[job-queue]==21.10

# ─── Scheduling ──────────────────────
apscheduler==3.10.4

# ─── LLM Client (OpenRouter = OpenAI-compatible) ──
openai==1.59.7

# ─── Google Drive (Knowledge Base) ───
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# ─── HTTP Client ─────────────────────
httpx==0.28.1

# ─── Utilities ───────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
